Create SerialEventCommander.ino
Used to fix the serial-menu for Mp3I2STeslaDriver.ino, I extruded the serial-driven menu allowing to type command and get an switch that receive an ExprCmd result. Belong to the enumeration of menu level, another variable of type stCmd_t as 12 member inside *TableCommand it's a method to inspect the TableCommand until you haven't found the last element CMD_END, every row are in the type stCmd_t holding two element, an enum of type ExprCmd, and a char* to hold the text typed by the user. 
Warning ! chChainCompare is not used but was intended  to compare a pair of 2 char* variable. To use it if you descrambled the string into 'unsigned char*' on your compilation option inside arduino-cli ( --build-property) do require to add "-DSTRINGCHAR_COMPARE", but actually not intended to be use.